webgl tests
issue with flashing clipRectangle

+++ lib/src/option/position.dart
@@ -57,7 +57,7 @@ class Position extends DisplayObjectContainer {
 
     });
 
-    return [x,y];
+    return [x.toInt(),y.toInt()];
   }
 
 }
